# docs/openapi.yaml
basePath: /api
definitions:
  handlers.CreateChallengeRequest:
    properties:
      category:
        type: string
      decay_threshold:
        type: integer
      description:
        type: string
      difficulty:
        type: string
      dynamic_scoring:
        type: boolean
      initial_points:
        type: integer
      minimum_points:
        type: integer
      name:
        type: string
      sql_dataset_url:
        type: string
      sql_enabled:
        type: boolean
      sql_schema_hint:
        type: string
      tags:
        type: string
      visible:
        type: boolean
    type: object
...
  models.Challenge:
    properties:
      category:
        type: string
      created_at:
        type: string
      decay_threshold:
        type: integer
      description:
        type: string
      difficulty:
        type: string
      dynamic_scoring:
        type: boolean
      file_url:
        type: string
      id:
        type: string
      initial_points:
        type: integer
      minimum_points:
        type: integer
      name:
        type: string
      sql_dataset_url:
        type: string
      sql_enabled:
        type: boolean
      sql_schema_hint:
        type: string
      tags:
        type: string
      updated_at:
        type: string
      visible:
        type: boolean
    type: object
...
  models.Question:
    properties:
      case_sensitive:
        type: boolean
      challenge_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      file_url:
        type: string
      flag_mask:
        type: string
      id:
        type: string
      name:
        type: string
      points:
        type: integer
      updated_at:
        type: string
    type: object
  models.ScoreboardEntry:
    properties:
      last_solve:
        type: string
      points:
        type: integer
      rank:
        type: integer
      solve_count:
        type: integer
      team_id:
        type: string
      team_name:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: Self-hosted CTF platform API. Most write endpoints require authentication
    via JWT cookie (auth_token).
  title: hCTF2 API
  version: 1.0.0
paths:
...
  /admin/challenges:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      - multipart/form-data
      parameters:
      - description: Challenge data
        in: body
        name: challenge
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateChallengeRequest'
      - description: Optional file attachment (max 50MB)
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Challenge'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Create a new challenge (admin only)
      tags:
      - Admin
  /admin/challenges/{id}:
    delete:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Empty response on success
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Delete a challenge (admin only)
      tags:
      - Admin
    put:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated challenge data
        in: body
        name: challenge
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateChallengeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Challenge updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Update an existing challenge (admin only)
      tags:
      - Admin
  /admin/challenges/{id}/file:
    delete:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragment indicating no file attached
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Delete a challenge's file attachment (admin only)
      tags:
      - Admin
...
  /api/ctftime:
    get:
      description: Returns team standings in the format expected by CTFtime.org. No
        authentication required.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              standings:
                items:
                  properties:
                    pos:
                      type: integer
                    score:
                      type: integer
                    team:
                      type: string
                  type: object
                type: array
            type: object
        "404":
          description: No team data available
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Export scoreboard in CTFtime.org JSON format
      tags:
      - Scoreboard
...
  /challenges:
    get:
      description: Returns all visible challenges. Admins also see hidden challenges.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Challenge'
            type: array
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: List all challenges
      tags:
      - Challenges
...
  /challenges/{id}:
    get:
      description: Returns challenge details and associated questions. Flags are hidden
        from non-admin users.
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              challenge:
                $ref: '#/definitions/models.Challenge'
              questions:
                items:
                  $ref: '#/definitions/models.Question'
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get a single challenge with its questions
      tags:
      - Challenges
...
  /scoreboard:
    get:
      description: Returns HTML table when called from HTMX (HX-Request header), otherwise
        returns JSON. Top 100 users.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ScoreboardEntry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get individual user scoreboard rankings
      tags:
      - Scoreboard
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"